Format only visible messages.
For this we store an offset from the last render in the
StatefulList. On rendering, we start with the offset (decreased to
selected if needed). We format all messages from the offset as long
they fit and up to the selected message (if any). Then we remove
all initial messages that did not fit (this happens when there is a
selected message outside).