Add cross compilation targets to CI and fix publish workflow. (#84)
Also fix the publish workflow to compile arm targets on nightly.
Nightly is required since signal-client is using the armv8 feature
in aes which is only available on nightly. We use the version of
nightly which is defined in the toolchain file of signal-client.

Another fix is not to use clang for native compilation. It fails
for arm cross compilation.