Add cursor tracking and multiline input navigation.
Cursor tracks a byte position in a text, and additionally a column
and line. It implements navigation of left/right,line up/down,
beginning/end of a line, and basic word operation. Also it supports
unicode and characters of width > 1.

Message input now supports switching between the single line and
multi-line modes via Alt+Enter. Previously, Alt+Enter inserted a new
line. Having a special mode for multi-line is more powerful, since it
allows navigation between lines with the same shortcuts as navigation
in channels and messages. And it allows pasting multi-line text.