Add cursor tracking and multiline input navigation (#131)
Cursor tracks a byte position in a text, and additionally a column
and line. It implements navigation of left/right,line up/down,
beginning/end of a line, and basic word operation. Also it supports
Unicode and characters of width > 1.

Message input now supports switching between the single line and
multi-line modes via Alt+Enter. Previously, Alt+Enter inserted a new
line. Having a special mode for multi-line is more powerful, since it
allows navigation between lines with the same shortcuts as navigation
in channels and messages. And it allows pasting multi-line text.

CI is switched to use the stable aarch64 rustc compiler. Nightly was
required due to signal-client using the armv8 feature in aes which
was only available on nightly.

Fixes #129