Sync contacts
Now we resolve user names from the synced address book, if possible.
We fallback to the profile name, if contact is not in the address
book. And eventually to contacts phone number, if profile retrieval
failed.

The contacts are synced on start-up of the app, however max one time
per hour.

Resolved #140