feat: add ephemeral status to sent messages on errors
Messages that could not be sent are marked as such by adding a read
field below showing the error message. The status is ephemeral and is
not stored. Resending of message is not implemented.

Also upgrade presage to incl. the fix which does not reuse websocket
connection for sending. This should fix sending messages.