feat: handle read notifications from other clients
The unread channel counters are now updated when a read notification is
recieved from another client.

Note that the counters are ephemeral and will be reset when the app is
restarted.

Marking a channel as read in Gurk does not update the unread counters in
other clients yet.

Related to #286