feat: add onboarding flow when there is no config (#391)
Onboarding consists of:

- displaying a welcome message
- asking for a display name
- asking for a passphrase
- asking where to store the passphrase

On macOS, additionally allow storing the passphrase in the keychain.
Also, attempt to load it from the keychain if no passphrase is provided
as a CLI argument or stored in the config file.

Additionally, phone number was removed from config. Also the default
config does not save paths anymore, if they have default values. In
particular, this makes the config portable.

Closes #370 Closes #360