fix: disable receipts (#408)
Currently, we have an unreproducible bug where contacts of a gurk user
get a message that a message could not be received from the user,
without user actually sending a message to the contact. The only silent
message we ever send is a receipt. To isolate the bug we disabled the
receipts handling.