fix: correct message selection behind date separators (#316, #324, #381) (#424)
This is a very hacky fix to the issue of incorrect message selection,
which affected reactions and copying (at least). `selected_message_id` was
making the presumption that indices in the UI list corresponded directly
to indices in the messages list. This presumption is incorrect when date
separators are used. In lieu of a proper architectural intervention, this
commit replicates the logic from `draw.rs`, checking whether a day
transition occurs in the local timezone that would prompt the insertion
of a date line - if so, an index correction is applied. Performance was
not profiled, but seems aesthetically unaffected.